requirements.txt
openai>=1.0.0
mcp>=1.20.0
mem0ai>=0.1.0
redis>=5.0.0
fastapi>=0.110.0
uvicorn>=0.30.0
pyyaml>=6.0
httpx>=0.27.0
